after hours of pissing around on the mac...gave up..so got this configured how i wanted in VS2015 now to see if it works on the mac
