--- src/Ocelot.ApiGateway/Infrastructure/Responses/Error.cs
@@ -1,12 +1,0 @@
-namespace Ocelot.ApiGateway.Infrastructure.Responses
-{
-    public abstract class Error 
-    {
-        public Error(string message)
-        {
-            Message = message;
-        }
-
-        public string Message { get; private set; }
-    }
-}
