--- src/Ocelot.ApiGateway/Infrastructure/Responses/ErrorResponse.cs
@@ -1,11 +1,0 @@
-namespace Ocelot.ApiGateway.Infrastructure.Responses
-{
-    using System.Collections.Generic;
-
-    public class ErrorResponse : Response
-    {
-        public ErrorResponse(List<Error> errors) : base(errors)
-        {
-        }
-    }
-}
