--- src/Ocelot.ApiGateway/Infrastructure/Responses/ErrorResponseGeneric.cs
@@ -1,11 +1,0 @@
-namespace Ocelot.ApiGateway.Infrastructure.Responses
-{
-    using System.Collections.Generic;
-
-    public class ErrorResponse<T> : Response<T>
-    {
-        public ErrorResponse(List<Error> errors) : base(errors)
-        {
-        }
-    }
-}
