--- src/Ocelot.ApiGateway/Infrastructure/Responses/OkResponse.cs
@@ -1,9 +1,0 @@
-namespace Ocelot.ApiGateway.Infrastructure.Responses
-{
-    public class OkResponse : Response
-    {
-        public OkResponse()
-        {
-        }
-    }
-}
