--- src/Ocelot.ApiGateway/Infrastructure/Responses/OkResponseGeneric.cs
@@ -1,9 +1,0 @@
-namespace Ocelot.ApiGateway.Infrastructure.Responses
-{
-    public class OkResponse<T> : Response<T>
-    {
-        public OkResponse(T data) : base(data)
-        {
-        }
-    }
-}
