--- src/Ocelot.ApiGateway/Infrastructure/Router/RouteKeyAlreadyExists.cs
@@ -1,12 +1,0 @@
-using Ocelot.ApiGateway.Infrastructure.Responses;
-
-namespace Ocelot.ApiGateway.Infrastructure.Router
-{
-    public class RouteKeyAlreadyExists : Error
-    {
-        public RouteKeyAlreadyExists(string message) 
-            : base(message)
-        {
-        }
-    }
-}
