--- src/Ocelot.ApiGateway/Infrastructure/Router/RouteKeyDoesNotExist.cs
@@ -1,12 +1,0 @@
-using Ocelot.ApiGateway.Infrastructure.Responses;
-
-namespace Ocelot.ApiGateway.Infrastructure.Router
-{
-    public class RouteKeyDoesNotExist : Error
-    {
-        public RouteKeyDoesNotExist(string message) 
-            : base(message)
-        {
-        }
-    }
-}
