+++ src/Ocelot.Library/Infrastructure/Responses/ErrorResponse.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Ocelot.Library.Infrastructure.Responses
+{
+    public class ErrorResponse : Response
+    {
+        public ErrorResponse(List<Error> errors) : base(errors)
+        {
+        }
+    }
+}
