+++ src/Ocelot.Library/Infrastructure/Responses/ErrorResponseGeneric.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Ocelot.Library.Infrastructure.Responses
+{
+    public class ErrorResponse<T> : Response<T>
+    {
+        public ErrorResponse(List<Error> errors) : base(errors)
+        {
+        }
+    }
+}
