+++ src/Ocelot.Library/Infrastructure/Responses/OkResponse.cs
@@ -1,0 +1,9 @@
+namespace Ocelot.Library.Infrastructure.Responses
+{
+    public class OkResponse : Response
+    {
+        public OkResponse()
+        {
+        }
+    }
+}
