+++ src/Ocelot.Library/Infrastructure/Responses/OkResponseGeneric.cs
@@ -1,0 +1,9 @@
+namespace Ocelot.Library.Infrastructure.Responses
+{
+    public class OkResponse<T> : Response<T>
+    {
+        public OkResponse(T data) : base(data)
+        {
+        }
+    }
+}
