+++ src/Ocelot.Library/Infrastructure/Responses/Response.cs
@@ -1,0 +1,19 @@
+using System.Collections.Generic;
+
+namespace Ocelot.Library.Infrastructure.Responses
+{
+    public abstract class Response
+    {
+        protected Response()
+        {
+            Errors = new List<Error>();
+        }
+
+        protected Response(List<Error> errors)
+        {
+            Errors = errors ?? new List<Error>();
+        } 
+
+        public List<Error> Errors { get; private set; }
+    }
+}
