+++ src/Ocelot.Library/Infrastructure/Router/RouteKeyAlreadyExists.cs
@@ -1,0 +1,12 @@
+using Ocelot.Library.Infrastructure.Responses;
+
+namespace Ocelot.Library.Infrastructure.Router
+{
+    public class RouteKeyAlreadyExists : Error
+    {
+        public RouteKeyAlreadyExists(string message) 
+            : base(message)
+        {
+        }
+    }
+}
