+++ src/Ocelot.Library/Infrastructure/Router/RouteKeyDoesNotExist.cs
@@ -1,0 +1,12 @@
+using Ocelot.Library.Infrastructure.Responses;
+
+namespace Ocelot.Library.Infrastructure.Router
+{
+    public class RouteKeyDoesNotExist : Error
+    {
+        public RouteKeyDoesNotExist(string message) 
+            : base(message)
+        {
+        }
+    }
+}
