+++ src/Ocelot.Library/Middleware/ProxyExtensions.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Builder;
+
+namespace Ocelot.Library.Middleware
+{
+    public static class ProxyExtensions
+    {
+        public static IApplicationBuilder UseProxy(this IApplicationBuilder builder)
+        {
+            return builder.UseMiddleware<ProxyMiddleware>();
+        }
+    }
+}
