+++ src/Ocelot.Library/Middleware/RequestLoggerExtensions.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Builder;
+
+namespace Ocelot.Library.Middleware
+{
+    public static class RequestLoggerExtensions
+    {
+        public static IApplicationBuilder UseRequestLogger(this IApplicationBuilder builder)
+        {
+            return builder.UseMiddleware<RequestLoggerMiddleware>();
+        }
+    }
+}
